Fix x64 pointer truncation in status-change vals (Safety Wall/Blade Stop/Dancing)
GCC's -Wpointer-to-int-cast/-Wint-to-pointer-cast flagged 38 size-changing
casts; triage found one real crash class (same as push_val): several skills
stash a live pointer in a status-change val, truncated to 32 bits on x64.

- skill.c/battle.c stored skill_unit_group*/block_list* as (int)sg / (int)target
  into sc_start4 vals; status_change_start took val1..4 as int and
  struct status_change_entry stored them as int -> high 32 bits lost.
- The readers cast it straight back to a full pointer and deref:
  battle.c:295 (SC_SAFETYWALL .val3), skill.c:9809 (SC_DANCING .val2),
  unit.c:823 (SC_BLADESTOP .val4) -> segfault on x64 the moment the skill
  fires (Safety Wall hit, Monk Blade Stop lock, Bard/Dancer ensemble).

Fix: widen val1..val4 to intptr_t in struct status_change_entry (map.h, the
live map-only struct; the wire/SQL struct status_change_data is untouched) and
in the status_change_start signature, and change the 7 store-site casts to
(intptr_t). The read sites are now clean intptr_t->pointer, no truncation.

Also fixed the format strings exposed by the widening / pre-existing on x64:
status.c:6873 and script.c:2290 (%d on intptr_t -> cast (int)), and
script.c:3388 mapreg SQL save used (intptr_t)data under %d -> (int)data to
match the TXT branch (was persisting via a mis-typed format).

The remaining 29 -Wpointer warnings are eAthena's benign "store a 32-bit int
in a void*" idiom (script var values, @users counters) plus offsetof/pointer-
difference math (atcommand.c:982, guild.c:1838) - value-preserving, not bugs.

Verified: make clean && make sql -> 0 errors, 0 intptr format warnings, all
10 status-change cast sites gone; full cluster (login+char+map) boots and
connects with the widened struct.

+++ src/map/battle.c
@@ -2861,11 +2861,11 @@ int battle_weapon_attack(struct block_list* src, struct block_list* target, unsi
 			int skilllv = tsc->data[SC_BLADESTOP_WAIT].val1;
 			int duration = skill_get_time2(MO_BLADESTOP,skilllv);
 			status_change_end(target, SC_BLADESTOP_WAIT, -1);
-			if(sc_start4(src, SC_BLADESTOP, 100, sd?pc_checkskill(sd, MO_BLADESTOP):5, 0, 0, (int)target, duration))
+			if(sc_start4(src, SC_BLADESTOP, 100, sd?pc_checkskill(sd, MO_BLADESTOP):5, 0, 0, (intptr_t)target, duration))
 		  	{	//Target locked.
 				clif_damage(src, target, tick, sstatus->amotion, 1, 0, 1, 0, 0); //Display MISS.
 				clif_bladestop(target,src,1);
-				sc_start4(target, SC_BLADESTOP, 100, skilllv, 0, 0,(int)src, duration);
+				sc_start4(target, SC_BLADESTOP, 100, skilllv, 0, 0,(intptr_t)src, duration);
 				return 0;
 			}
 		}
